pdf2text 2/requirements.txt
# Core PDF processing
pdfplumber==0.11.0
PyPDF2==3.0.1
pdf2docx==0.5.8

# AI API integration
requests==2.31.0
openai==1.12.0

# Document generation
python-docx==1.1.0
markdown==3.5.2

# Web framework
streamlit==1.31.0
streamlit-extras==0.3.6

# Data processing
pandas==2.2.0
numpy==1.26.4
openpyxl==3.1.2

# Async processing
aiohttp==3.9.3

# Utilities
python-dotenv==1.0.1
Pillow==10.2.0
tqdm==4.66.1

# Logging
loguru==0.7.2
